feat: tree diff algorithm
Implements a merkle tree difference algorithm to identify inconsistent
ranges between a "local" tree, and a "peer" tree. Returns the ranges
that should be pulled from the "peer" to converge the local tree.

This implementation performs a depth-first walk of the PageRange list
(itself a serialised tree!) and optimistically assumes page range bounds
that overlap but do not match are the only cause of the page
inconsistency, marking only the key range difference as inconsistent.
This optimises for monotonic keys, which are recommended by the MST
paper to minimise page invalidations for insert operations.

This algorithm terminates if both trees perform sync rounds - if one
tree never pulls inconsistent keys from the other, then the latter may
get "stuck" always marking page ranges as inconsistent but unable to
fully converge.