perf: less branching in page serialisation
When serialising a tree, the root may be empty if no values have been
added. This must be handled as a special case, as no other pages can
ever be empty.

This change checks the empty root special case once at the start,
eliminating a branch in the visitor that's hit for every page.