perf: reduce keys fetched to converge
Change the diff evaluation to continue evaluating within subtrees,
instead of prematurely ending sub-tree evaluation and marking the
remaining pages as inconsistent.

This leads to more pages being marked as consistent, causing fewer keys
to be read when converging - tests show ~26% reduction in keys read when
converging a set of reference trees.