perf: disable monotonic key optimisation
Removes the optimisation for monotonic key inserts into the MST.

In insert-only trees with monotonic keys, this optimisation can vastly
reduce the number of keys fetched during syncs. However, it's not free -
the trade-off is increased sync overhead for entry updates and/or
non-monotonic keys, which is a vastly more common workload. Therefore
this optimisation has been disabled - it may return as an option in the
future.

This reduces the average number of sync rounds needed to converge the
reference trees by ~half, and the number of keys exchanged by ~50%,
making syncs much cheaper for the average case.